Web: Single Protein: User Selects Mods Displayed. Open Mods Displayed
On Single Protein for Single Search and Multiple Searches (NOT
Experiment)
The User can select which Modification Masses are displayed in the
Peptide List in the Generated Peptide Strings (new for Single Search).
New for Generated Peptide Strings is display of Open Modifications
(rounded to whole number)

+++ limelight_webapp/src/main/java/org/yeastrc/limelight/limelight_webapp/spring_mvc_parts/data_pages/rest_controllers/single_project_search_id/OpenModifications_PerReportedPeptide_For_ReportedPeptideIds_Single_ProjSearchID_RestWebserviceController.java
@@ -262,7 +262,7 @@ public class OpenModifications_PerReportedPeptide_For_ReportedPeptideIds_Single_
     public static class WebserviceResult_Item {
 
     	private int reportedPeptideId;
-    	private double mass;
+    	private int mass;
 
     	public WebserviceResult_Item() {}
 
@@ -278,11 +278,11 @@ public class OpenModifications_PerReportedPeptide_For_ReportedPeptideIds_Single_
     		this.reportedPeptideId = reportedPeptideId;
     	}
 
-    	public double getMass() {
+    	public int getMass() {
     		return mass;
     	}
 
-    	public void setMass(double mass) {
+    	public void setMass(int mass) {
     		this.mass = mass;
     	}
     }
